% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Returns the entropy of a multivariate normal distribution with a specified
covariance matrix}
\usage{
entropy(x)
}
\arguments{
\item{x}{A k x k covariance matrix}
}
\value{
The entropy of the corresponding multivariate normal distribution
}
\description{
Returns the entropy of a multivariate normal distribution with a specified
covariance matrix
}
