% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy_omega}
\alias{entropy_omega}
\title{Calculate the entropy for a k x k correlation matrix with a constant
(positive) correlation}
\usage{
entropy_omega(r, k)
}
\arguments{
\item{r}{The correlation coefficient between all pairs of variables}

\item{k}{The dimension of the correlation matrix}
}
\value{
The entropy of the corresponding multivariate normal distribution
}
\description{
Calculate the entropy for a k x k correlation matrix with a constant
(positive) correlation
}
