% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eigen.R
\name{get_eigen}
\alias{get_eigen}
\title{Return eigenvalue statistics from a set of posterior correlation
matrices}
\usage{
get_eigen(Omega)
}
\arguments{
\item{Omega}{An array of correlation matrices, typically from get_Omega().}
}
\value{
A data frame with the eigenvalues, lead eigenvalue, and variance
of eigenvalues for each of the correlation matrices
}
\description{
This function takes an n_iter x n_traits x n_traits set of correlation
matrices and returns a data frame. The first n_trait columns of the data
frame are the eigenvalues (prefixed with "EV_"). The remaining two are
the lead eigenvalue (Lead_EV), and the variance of eigenvalues (Var_EV).
}
