% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_ent_R}
\alias{get_ent_R}
\title{Calculate the correlation coefficient a multivariate normal with all
pairwise correlations equal that has an entropy equivalent to the
observed entropy.}
\usage{
get_ent_R(Omega, tol = 1e-14)
}
\arguments{
\item{Omega}{An n_iter x n_traits x n_traits array of correlation matrices}

\item{tol}{Stop the search when the entropy difference is less than tol}
}
\value{
A vector of r equivalents, i.e., the pairwise corrleation in a
constant correlation matrix producing entropies equivalent to those in
the vector of entropies supplied
}
\description{
Calculate the correlation coefficient a multivariate normal with all
pairwise correlations equal that has an entropy equivalent to the
observed entropy.
}
