% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Returns the entropies for a set of covariance matrices}
\usage{
get_entropy(Omega)
}
\arguments{
\item{Omega}{An n_iter x n_traits x n_traits array of covariance matrices}
}
\value{
A numeric vector containing the entropy associated with each
covariance matrix
}
\description{
Returns the entropies for a set of covariance matrices
}
