% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Omega.R
\name{get_Omega}
\alias{get_Omega}
\title{Extract array of correlation matrices from output of brm()}
\usage{
get_Omega(output_model)
}
\arguments{
\item{output_model}{An object of class brmsfit}
}
\value{
An n_iter x trait x trait array
}
\description{
This function takes an object of class brmsfit produced by fitting a
multivariate response model, extracts the correlations, checks that all of
the correlation matrices are positive definite (printing a warning for any
that aren't), and returns them.
}
