% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Omega.R
\name{get_Omega_boot}
\alias{get_Omega_boot}
\title{Construct an array of correlation matrices from a bootstrap sample of
correlation matrices.}
\usage{
get_Omega_boot(sample)
}
\arguments{
\item{sample}{The bootstrap sample}
}
\value{
An n_iter x trait x trait array
}
\description{
This function takes an matrix consisting of bootstrap samples of a
correlation matrix with each row consisting of a bootstrap sample with
for one element of the correlation matrix. Note: The number of rows
in the matrix must be a perfect square, and the square root of the number
of rows must be equal to the number of traits.
}
