% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_statistics.R
\name{get_summary}
\alias{get_summary}
\title{Return posterior mean and specified quantiles for one of the trait
integration measures}
\usage{
get_summary(result_vector, probs)
}
\arguments{
\item{result_vector}{The vector of sampling results for a trait integration
measure}

\item{probs}{Numeric vector of probabilities for the quantiles}
}
\value{
A named list with components Mean (corresponding to the posterior
mean) and quantile (a vector with components corresponding to the
specified probabilities
}
\description{
Return posterior mean and specified quantiles for one of the trait
integration measures
}
