% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_graph.R
\name{plot_posterior_correlation_boot}
\alias{plot_posterior_correlation_boot}
\title{Visualize a the posterior distribution of a correlation matrix from
a bootstrap sample of correlation matrices}
\usage{
plot_posterior_correlation_boot(
  Omega,
  labels = NULL,
  style = "arc",
  palette = "RdBu",
  label_edges = FALSE,
  digits = 2
)
}
\arguments{
\item{Omega}{A n_sample x n_trait x n_trait array of correlation matrices}

\item{labels}{Labels for the nodes. If `NULL` then the names are the
columnnames of Omega}

\item{style}{Style of visualization, "arc" (default) or "line"}

\item{palette}{RColorBrewer palette to use. See `apply_colors()` for
details. Default: "RdBu"}

\item{label_edges}{Label edges of graph with posterior mean correlation.
Default: FALSE}

\item{digits}{Number of digits to use in rounding posterior mean correlation
(only relevant if label_edges == TRUE). Default: 2}
}
\value{
A list with two elements

`p` A gggraph object that can be further modified

`R` The posterior mean correlation matrix
}
\description{
This function plots the posterior correlation matrix from a bootstrapped
sample of correlation matrices. It consists ofa series of arcs with the
color of each connection representing the posterior mean of the correlation
coefficient and the width representing the one-sided probability that the
coefficient is different from 0.
}
