% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_graph.R
\name{plotter_for_posterior_correlation}
\alias{plotter_for_posterior_correlation}
\title{Internal function to make plot}
\usage{
plotter_for_posterior_correlation(
  Omega,
  labels,
  style,
  palette,
  label_edges,
  digits
)
}
\arguments{
\item{Omega}{A n_sample x n_trait x n_trait array of correlation matrices}

\item{labels}{Labels for the nodes}

\item{style}{Style of visualization}

\item{palette}{RColorBrewer palette to use}

\item{label_edges}{Label edges of graph with posterior mean correlation}

\item{digits}{Number of digits to use}
}
\value{
A list with two elements

`p` A ggraph object

`R` The posterior mean correlation matrix
}
\description{
This function is used internally to plot a representation of the
posterior correlation matrix. It is called by either
`plot_posterior_correlation()` for analyses using `brms` or
`plot_posterior_correlation_boot()` for analyses using a hand-coded
bootstrap sample.
}
