% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traits.R
\name{report_trait_names}
\alias{report_trait_names}
\title{Report the traits found in the model}
\usage{
report_trait_names(output_model)
}
\arguments{
\item{output_model}{An object of class `brmsfit`}
}
\value{
A vector of parameter names
}
\description{
This is a utility function to make it easy to find the traits used in a
`brms` analysis and to report them in the order they will appear when
the posterior correlation matrix is plotted
}
