% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_statistics.R
\name{summarize_statistics}
\alias{summarize_statistics}
\title{Summarize results returned by `assemble_statistics()`}
\usage{
summarize_statistics(results, probs = c(0.025, 0.975), digits = 3)
}
\arguments{
\item{results}{A data fram returned by assemble statistics}

\item{probs}{Numeric vector of probabilities for quantiles. Default:
c(0.025, 0.975)}

\item{digits}{Number of digits to report in results. Default: 3}
}
\description{
This function provides a quick summary of results returned by
`assemble_statistics`. The summary includes the posterior mean and
quantiles for the leading eigenvalue, the variance of eigenvalues, and
entR. By default only the 2.5% and 97.5% quantiles are reported, but
any numeric vector of probabilities in [0, 1] can be specified. See the
documentation for `quantile()` for additional details.
}
\examples{
tmp <- assemble_statistics(mod_wild)
summarize_statistics(tmp)

}
