% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_graph.R
\name{apply_colors}
\alias{apply_colors}
\title{Apply colors to a correlation matrix using ColorBrewer palettes}
\usage{
apply_colors(R, name = "RdBu", viridis = FALSE)
}
\arguments{
\item{R}{A correlation matrix}

\item{name}{The name of a ColorBrewer palette. This should be a diverging
palette with at least 9 distinct colors. All diverging ColorBrewer palettes
meet this requirement. Default: "RdBu"}

\item{viridis}{Use a Viridis palette. If true, the turbo palette is used.}
}
\value{
A data frame with columns \code{color} and \code{label}. \code{label} is the
correlation coefficient.
}
\description{
This function is used internally to (a) translate correlation coefficients
to colors on a (diverging) ColorBrewer palette and (b) return both the
color and the corresponding correlation coefficient.
}
