% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_statistics.R
\name{assemble_statistics_boot}
\alias{assemble_statistics_boot}
\title{Assemble eigenvalue and entR statistics from bootstrap samples}
\usage{
assemble_statistics_boot(Omega)
}
\arguments{
\item{sample}{A bootstrap sample of correlation matrices. Each row
is the bootstrap sample for one element of the correlation matrix.}
}
\value{
A data frame with the eigenvalues, lead eigenvalue, variance
of eigenvalues, and r_eff statistic for each of the correlation matrices
in the bootstrap sample
}
\description{
This function encapsulates all of the computations associated with
assessing phenotypic integration using a bootstrap sample of correlation
matrices. See `vignette("Bootstrapping", package = "traitintegration")`
for a complete example.
}
